+++ tests/testthat/test-active-filter.R
@@ -19,7 +19,9 @@ with_mock_HTTP({
         expect_identical(activeFilter(ds2), ds$gender == "Male")
     })
     test_that("Active filter persists on refreshing dataset", {
-        expect_identical(activeFilter(refresh(ds2)), ds$gender == "Male")
+        ## Compare @expressions because of a quirk of the mock backend
+        expect_identical(activeFilter(refresh(ds2))@expression,
+            (ds$gender == "Male")@expression)
     })
     test_that("Further [ on a filtered dataset ands the filters together", {
         expect_identical(activeFilter(ds3),
@@ -66,13 +68,29 @@ with_mock_HTTP({
     })
 
     test_that("activeFilter from CrunchExpr", {
-        ## Need the @expression here because CrunchExpr@filter is list
         expect_identical(activeFilter((ds$birthyr - ds$starttime)[ds$gender == "Male"]),
-            (ds$gender == "Male")@expression)
+            ds$gender == "Male")
     })
     test_that("activeFilter passes across operations among vars/exprs", {
-        skip("TODO: filtered variables/exprs should pass their filters along")
         expect_identical(activeFilter(ds$birthyr[ds$gender == "Male"] - ds$starttime[ds$gender == "Male"]),
+            ds$gender == "Male")
+    })
+    test_that("Vars with different activeFilters can't combine", {
+        expect_error(ds$birthyr[ds$gender == "Male"] - ds$starttime[ds$gender == "Female"],
+            "Cannot combine expressions with different filters")
+    })
+    test_that("Exprs with different activeFilters can't combine", {
+        expect_error((ds$birthyr - ds$starttime)[ds$gender == "Male"] - (ds$birthyr - ds$starttime)[ds$gender == "Female"],
+            "Cannot combine expressions with different filters")
+    })
+    test_that("Vars/exprs together with different active filters can't combine", {
+        expect_error(ds$birthyr[ds$gender == "Male"] - (ds$birthyr - ds$starttime)[ds$gender == "Female"],
+            "Cannot combine expressions with different filters")
+    })
+    test_that("Can combine var with activeFilter with non-Crunch object", {
+        expect_identical(activeFilter(ds$birthyr[ds$gender == "Male"] - 50),
+            ds$gender == "Male")
+        expect_identical(activeFilter(2016 - ds$birthyr[ds$gender == "Male"]),
             ds$gender == "Male")
     })
 })